+++ finance/src/test/kotlin/net/corda/flows/IssuerFlowTest.kt
@@ -9,6 +9,7 @@ import net.corda.core.flows.FlowException
 import net.corda.core.flows.FlowStateMachine
 import net.corda.core.getOrThrow
 import net.corda.core.map
+import net.corda.core.node.services.deanonymiseParty
 import net.corda.core.serialization.OpaqueBytes
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.DUMMY_NOTARY
@@ -54,11 +55,12 @@ class IssuerFlowTest {
     }
 
     private fun runIssuerAndIssueRequester(amount: Amount<Currency>, issueToPartyAndRef: PartyAndReference) : RunResult {
+        val resolvedIssuerParty = bankOfCordaNode.services.identityService.deanonymiseParty(issueToPartyAndRef) ?: throw IllegalStateException()
         val issuerFuture = bankOfCordaNode.initiateSingleShotFlow(IssuerFlow.IssuanceRequester::class) {
-            otherParty -> IssuerFlow.Issuer(issueToPartyAndRef.party)
+            otherParty -> IssuerFlow.Issuer(resolvedIssuerParty)
         }.map { it.stateMachine }
 
-        val issueRequest = IssuanceRequester(amount, issueToPartyAndRef.party, issueToPartyAndRef.reference, bankOfCordaNode.info.legalIdentity)
+        val issueRequest = IssuanceRequester(amount, resolvedIssuerParty, issueToPartyAndRef.reference, bankOfCordaNode.info.legalIdentity)
         val issueRequestResultFuture = bankClientNode.services.startFlow(issueRequest).resultFuture
 
         return IssuerFlowTest.RunResult(issuerFuture, issueRequestResultFuture)