+++ docs/source/release-notes.rst
@@ -3,13 +3,12 @@ Release notes
 
 Here are brief summaries of what's changed between each snapshot release.
 
-Milestone 8
+Milestone 9
 -----------
 
 * API:
 
-    * ``Party`` equality is now based on the owning key, rather than the owning key and name. This is important for
-      party anonymisation to work, as each key must identify exactly one party.
+    * Pseudonymous ``AnonymousParty`` class added as a superclass of ``Party``.
 
 Milestone 7
 -----------